1.StatMod {statmod}	R Documentation

Introduction to the StatMod Package

Description

This library packages together those functions, other than those for microarray data analysis, which I wish to make public. A change-log for this package is available from http://www.statsci.org/r/changelog.txt.

Contributions to this library have also been made by Paul Bagshaw, Centre National d'Etudes des Telecommunications (DIH/DIPS), France (qinvgauss), and Trevor Park, Department of Statistics, University of Florida (rinvgauss).

Generalized Linear Models

tweedie, canonic.digamma, unitdeviance.digamma, varfun.digamma, cumulant.digamma, d2cumulant.digamma, meanval.digamma and logmdigamma are functions to fit non-standard generalized linear models related to the gamma distribution.

qres implements randomized quantile residuals for generalized linear models.

Growth Curves

compareGrowthCurves, compareTwoGrowthCurves and meanT are functions to test for differences between growth curves with repeated measurements on subjects.

Limiting Dilution Analysis

limdil implements limiting dilution analysis using complemenary log-log binomial generalized linear model regression, with some improvements on previous programs.

Probability Distributions

qinvgauss, dinvgauss, pinvgauss and rinvgauss perform probability calculations for the inverse Gaussian distribution.

gauss.quad and gauss.quad.prob compute Gaussian Quadrature with probability distributions.

Tests

hommel.test performs Hommel's multiple comparison tests.

power.fisher.test computes the power of Fisher's Exact Test for comparing proportions.

sage.test is a fast approximation to Fisher's exact test for each tag for comparing two Serial Analysis of Gene Expression (SAGE) libraries.

Variance Models

mixedModel2, mixedModel2Fit and glmgam.fit fit mixed linear models.

remlscore and remlscoregamma fit heteroscedastic and varying dispersion models by REML. welding is an example data set.

Matrix Computations

matvec and vecmat facilitate multiplying matrices by vectors.

Author(s)

Gordon Smyth

[Package statmod version 1.3.8 Index]