deprecated {statmod}	R Documentation

Deprecated Functions in statmod Package

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Usage


randomizedBlock(formula, random, weights=NULL, only.varcomp=FALSE, data=list(), subset=NULL, contrasts=NULL, tol=1e-6, maxit=50, trace=FALSE)
randomizedBlockFit(y, X, Z, w=NULL, only.varcomp=FALSE, tol=1e-6, maxit=50, trace=FALSE)

Arguments

The arguments formula, weights, data, subset and contrasts have the same meaning as in lm. The arguments y, X and w have the same meaning as in lm.wfit.

formula 	 formula specifying the fixed model.

random 	 vector or factor specifying the blocks corresponding to random effects.

weights 	 optional vector of prior weights.

only.varcomp 	 logical value, if TRUE computation of standard errors and fixed effect coefficients will be skipped

data 	 an optional data frame containing the variables in the model.

subset 	 an optional vector specifying a subset of observations to be used in the fitting process.

contrasts 	 an optional list. See the contrasts.arg argument of model.matrix.default.

tol 	 small positive numeric tolerance, passed to glmgam.fit

maxit 	 maximum number of iterations permitted, passed to glmgam.fit

trace 	 logical value, passed to glmgam.fit. If TRUE then working estimates will be printed at each iteration.

y 	 numeric response vector

X 	 numeric design matrix for fixed model

Z 	 numeric design matrix for random effects

w 	 optional vector of prior weights

Details

randomizedBlock and randomizedBlockFit have been renamed to mixedModel2 and mixedModel2Fit respectively.

[Package statmod version 1.3.8 Index]