forward {statmod}	R Documentation

Forward Selection of Covariates for Multiple Regression

Description

Fit a multi-group negative-binomial model to SAGE data, with Pearson estimation of the common overdispersion parameter.

Usage

forward(y, x, xkept=NULL, intercept=TRUE, nvar=ncol(x))

Arguments

y
numeric response vector.

x
numeric matrix of covariates, candidates to be added to the regression.

xkept
numeric matrix of covariates to be included in the starting regression.

intercept
logical, should an intercept be added to xkept?

nvar
integer, number of covariates from x to add to the regression.

Details

This function has the advantage that x can have many more columns than the length of y.

Value

Integer vector of length nvar, giving the order in which columns of x are added to the regression.

Author(s)

Gordon Smyth

See Also

step

Examples

y <- rnorm(10)
x <- matrix(rnorm(10*5),10,5)
forward(y,x)

[Package statmod version 1.5.1 Index]