glmgam.fit {statmod}	R Documentation

Fit Gamma Generalized Linear Model by Fisher Scoring with Identity Link

Description

Fit a generalized linear model with secure convergence.

Usage

glmgam.fit(X, y, coef.start = NULL, tol = 1e-6, maxit = 50, trace = FALSE)

Arguments

X
design matrix, assumed to be of full column rank. Missing values not allowed.

y
numeric vector of responses. Negative or missing values not allowed.

coef.start
numeric vector of starting values for the regression coefficients

tol
small positive numeric value giving convergence tolerance

maxit
maximum number of iterations allowed

trace
logical value. If TRUE then output diagnostic information at each iteration.

Details

This function implements a modified Fisher scoring algorithm for generalized linear models, similar to the Levenberg-Marquardt algorithm for nonlinear least squares. The Levenberg-Marquardt modification checks for a reduction in the deviance at each step, and avoids the possibility of divergence. The result is a very secure algorithm that converges for almost all datasets.

glmgam.fit is in principle equivalent to glm.fit(X,y,family=Gamma(link="identity")) but with much more secure convergence.

Value

List with the following components:

coefficients
numeric vector of regression coefficients

fitted
numeric vector of fitted values

deviance
residual deviance

iter
number of iterations used to convergence. If convergence was not achieved then iter is set to maxit+1.

Author(s)

Gordon Smyth and Yunshun Chen

References

Dunn, PK, and Smyth, GK (2018). Generalized linear models with examples in R. Springer, New York, NY. doi:10.1007/978-1-4419-0118-7

See Also

glmgam.fit is called by mixedModel2Fit.

glm is the standard glm fitting function in the stats package.

Examples

y <- rgamma(10, shape=5)
X <- cbind(1, 1:10)
fit <- glmgam.fit(X, y, trace=TRUE)

[Package statmod version 1.5.1 Index]