mixedModel2 {statmod}	R Documentation

Fit Mixed Linear Model with 2 Error Components

Description

Fits a mixed linear model by REML. The linear model contains one random factor apart from the unit errors.

Usage

mixedModel2(formula, random, weights=NULL, only.varcomp=FALSE, data=list(),
            subset=NULL, contrasts=NULL, tol=1e-6, maxit=50, trace=FALSE)
mixedModel2Fit(y, X, Z, w=NULL, only.varcomp=FALSE, tol=1e-6, maxit=50, trace=FALSE)
randomizedBlock(formula, random, weights=NULL, only.varcomp=FALSE, data=list(),
            subset=NULL, contrasts=NULL, tol=1e-6, maxit=50, trace=FALSE)
randomizedBlockFit(y, X, Z, w=NULL, only.varcomp=FALSE, tol=1e-6, maxit=50, trace=FALSE)

Arguments

The arguments formula, weights, data, subset and contrasts have the same meaning as in lm. The arguments y, X and w have the same meaning as in lm.wfit.

formula
formula specifying the fixed model.

random
vector or factor specifying the blocks corresponding to random effects.

weights
optional vector of prior weights.

only.varcomp
logical value, if TRUE computation of standard errors and fixed effect coefficients will be skipped

data
an optional data frame containing the variables in the model.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

contrasts
an optional list. See the contrasts.arg argument of model.matrix.default.

tol
small positive numeric tolerance, passed to glmgam.fit

maxit
maximum number of iterations permitted, passed to glmgam.fit

trace
logical value, passed to glmgam.fit. If TRUE then working estimates will be printed at each iteration.

y
numeric response vector

X
numeric design matrix for fixed model

Z
numeric design matrix for random effects

w
optional vector of prior weights

Details

Note that randomizedBlock and mixedModel2 are alternative names for the same function.

This function fits the model y=Xb+Zu+e where b is a vector of fixed coefficients and u is a vector of random effects. Write n for the length of y and q for the length of u. The random effect vector u is assumed to be normal, mean zero, with covariance matrix σ^2_uI_q while e is normal, mean zero, with covariance matrix σ^2I_n. If Z is an indicator matrix, then this model corresponds to a randomized block experiment. The model is fitted using an eigenvalue decomposition that transforms the problem into a Gamma generalized linear model. To the knowledge of the author, this is an original and unpublished approach to the problem of fitting mixed models.

Note that the block variance component varcomp[2] is not constrained to be non-negative. It may take negative values corresponding to negative intra-block correlations. However the correlation varcomp[2]/sum(varcomp) must lie between -1 and 1.

Missing values in the data are not allowed.

This function is in principle equivalent to lme(fixed=formula,random=~1|random) except that the block variance component is not constrained to be non-negative. If the block variance component is non-negative, then this function is faster and more accurate than lme for small to moderate size data sets but slower than lme when the number of observations is large.

This function tends to be fast and reliable, compared to competitor functions that fit randomized block models, when then number of observations is small, say no more than 200. However it becomes quadratically slow as the number of observations increases because of the need to compute two singular value decompositions of order nearly equal to the number of observations, although this can be limited to only one decomposition if only.varcomp = TRUE). For these reasons, this function is a good choice when fitting large numbers of mixed models to small datasets but is not optimal as currently implemented for fitting mixed models to very large data sets.

Value

A list with the components:

varcomp
numeric vector of length two containing the residual and block components of variance.

se.varcomp
standard errors for the variance components (if only.varcomp=FALSE).

coefficients
numeric vector of fixed coefficients (if only.varcomp=FALSE).

se.coefficients
standard errors for the fixed coefficients (if only.varcomp=FALSE).

Author(s)

Gordon Smyth

References

Venables, W., and Ripley, B. (2002). Modern Applied Statistics with S-Plus, Springer.

See Also

glmgam.fit, lme, lm, lm.fit

Examples

#  Compare with first data example from Venable and Ripley (2002),
#  Chapter 10, "Linear Models"
library(MASS)
data(petrol)
out <- mixedModel2(Y~SG+VP+V10+EP, random=No, data=petrol)
cbind(varcomp=out$varcomp,se=out$se.varcomp)

[Package statmod version 1.4.36 Index]