tweedie {statmod}	R Documentation

Tweedie Generalized Linear Models

Description

Produces a generalized linear model family object with any power variance function and any power link. Includes the Gaussian, Poisson, gamma and inverse-Gaussian families as special cases.

Usage

tweedie(var.power=0, link.power=1-var.power)

Arguments

var.power
index of power variance function

link.power
index of power link function. link.power=0 produces a log-link. Defaults to the canonical link, which is 1-var.power.

Details

This function provides access to a range of generalized linear model response distributions which are not otherwise provided by R, or any other package for that matter. It is also useful for accessing distribution/link combinations which are disallowed by the R glm function.

Let μ_i = E(y_i) be the expectation of the ith response. We assume that

μ_i^q = x_i^Tb, var(y_i) = φ μ_i^p

where x_i is a vector of covariates and b is a vector of regression cofficients, for some φ, p and q. This family is specified by var.power = p and link.power = q. A value of zero for q is interpreted as \log(μ_i) = x_i^Tb.

The variance power p characterizes the distribution of the responses y. The following are some special cases:

p 	 Response distribution

0 	 Normal

1 	 Poisson

(1, 2) 	 Compound Poisson, non-negative with mass at zero

2 	 Gamma

3 	 Inverse-Gaussian

> 2 	 Stable, with support on the positive reals

The name Tweedie has been associated with this family by Joergensen (1987) in honour of M. C. K. Tweedie.

Value

A family object, which is a list of functions and expressions used by glm and gam in their iteratively reweighted least-squares algorithms. See family and glm in the R base help for details.

Author(s)

Gordon Smyth

References

Tweedie, M. C. K. (1984). An index which distinguishes between some important exponential families. In Statistics: Applications and New Directions. Proceedings of the Indian Statistical Institute Golden Jubilee International Conference. (Eds. J. K. Ghosh and J. Roy), pp. 579-604. Calcutta: Indian Statistical Institute.

Joergensen, B. (1987). Exponential dispersion models. J. R. Statist. Soc. B 49, 127-162.

Smyth, G. K. (1996). Regression modelling of quantity data with exact zeroes. Proceedings of the Second Australia-Japan Workshop on Stochastic Models in Engineering, Technology and Management. Technology Management Centre, University of Queensland, pp. 572-580.

Joergensen, B. (1997). Theory of Dispersion Models, Chapman and Hall, London.

Smyth, G. K., and Verbyla, A. P., (1999). Adjusted likelihood methods for modelling dispersion in generalized linear models. Environmetrics 10, 695-709.

See Also

glm, family, dtweedie

Examples

y <- rgamma(20,shape=5)
x <- 1:20
# Fit a poisson generalized linear model with identity link
glm(y~x,family=tweedie(var.power=1,link.power=1))

# Fit an inverse-Gaussion glm with log-link
glm(y~x,family=tweedie(var.power=3,link.power=0)) 

[Package statmod version 1.4.32 Index]