welding {statmod}	R Documentation

Data: Tensile Strength of Welds

Description

This is a highly fractionated two-level factorial design employed as a screening design in an off-line welding experiment performed by the National Railway Corporation of Japan. There were 16 runs and 9 experimental factors. The response variable is the observed tensile strength of the weld, one of several quality characteristics measured. All other variables are at plus and minus levels.

Usage

data(welding)

Format

A data frame containing the following variables. All the explanatory variables are numeric with two levels, -1 and 1.

Variable 	 Description

Rods 	 Kind of welding rods

Drying 	 Period of drying

Material 	 Welded material

Thickness 	 Thickness

Angle 	 Angle

Opening 	 Opening

Current 	 Current

Method 	 Welding method

Preheating 	 Preheating

Strength 	 Tensile strength of the weld in kg/mm. The response variable.

Source

http://www.statsci.org/data/general/welding.html

References

Smyth, G. K., Huele, F., and Verbyla, A. P. (2001). Exact and approximate REML for heteroscedastic regression. Statistical Modelling 1, 161-175.

Smyth, G. K. (2002). An efficient algorithm for REML in heteroscedastic regression. Journal of Computational and Graphical Statistics 11, 1-12.

[Package statmod version 1.4.36 Index]