fitNBP {statmod}	R Documentation

Negative Binomial Model for SAGE Libraries with Pearson Estimation of Dispersion

Description

Fit a multi-group negative-binomial model to SAGE data, with Pearson estimation of the common overdispersion parameter.

Usage


fitNBP(y,group=NULL,lib.size=colSums(y),verbose=FALSE)

Arguments

y 	 numeric matrix giving counts. Rows correspond to tags (genes) and columns to SAGE libraries.

group 	 factor indicating which library belongs to each group. If NULL then one group is assumed.

lib.size 	 vector giving total number of tags in each library.

verbose 	 logical, if TRUE then iteration progress information is output.

Details

The overdispersion parameter is estimated equating the Pearson goodness of fit to its expectation. The variance is assumed to be of the form Var(y)=mu*(1+phi*mu) where E(y)=mu and phi is the dispersion parameter. All tags are assumed to share the same dispersion.

For given dispersion, the model for each tag is a negative-binomial generalized linear model with log-link and log(lib.size) as offset. The coefficient parametrization used is that corresponding to the formula ~0+group+offset(log(lib.size).

Except for the dispersion being common rather than genewise, the model fitted by this function is equivalent to that proposed by Lu et al (2005). The numeric algorithm used is that of alternating iterations (Smyth, 1996) using Newton's method as the outer iteration for the dispersion parameter starting at phi=0. This iteration is monotonically convergent for the dispersion.

Value

List with components

coefficients 	 numeric matrix of rates for each tag (gene) and each group

fitted.values 	 numeric matrix of fitted values

dispersion 	 estimated dispersion parameter

Author(s)

Gordon Smyth

References

Lu, J, Tomfohr, JK, Kepler, TB (2005). Identifying differential expression in multiple SAGE libraries: an overdispersed log-linear model approach. BMC Bioinformatics 6,165.

Smyth, G. K. (1996). Partitioned algorithms for maximum likelihood and other nonlinear estimation. Statistics and Computing, 6, 201-216.

See Also

sage.test

Examples


# True value for dispersion is 1/size=2/3
# Note the Pearson method tends to under-estimate the dispersion
y <- matrix(rnbinom(10*4,mu=4,size=1.5),10,4)
lib.size <- rep(50000,4)
group <- c(1,1,2,2)
fit <- fitNBP(y,group=group,lib.size=lib.size)
logratio <- fit$coef %*% c(-1,1)

[Package statmod version 1.3.8 Index]