invgauss {statmod}	R Documentation

Inverse Gaussian Distribution

Description

Density, cumulative probability, quantiles and random generation for the inverse Gaussian distribution.

Usage


dinvgauss(x, mu, lambda=1)
pinvgauss(q, mu, lambda=1)
qinvgauss(p, mu, lambda=1)
rinvgauss(n, mu, lambda=1)

Arguments

x 	 vector of quantiles. Missing values (NAs) are allowed.

q 	 vector of quantiles. Missing values (NAs) are allowed.

p 	 vector of probabilities. Missing values (NAs) are allowed.

n 	 sample size. If length(n) is larger than 1, then length(n) random values are returned.

mu 	 vector of (positive) means. This is replicated to be the same length as p or q or the number of deviates generated.

lambda 	 vector of (positive) precision parameters. This is replicated to be the same length as p or q or the number of deviates generated.

Details

The inverse Gaussian distribution takes values on the positive real line. The variance of the distribution is $μ^3/λ$. Applications of the inverse Gaussian include sequential analysis, diffusion processes and radiotechniques. The inverse Gaussian is one of the response distributions used in generalized linear models.

Value

Vector of same length as x or q giving the density (dinvgauss), probability (pinvgauss), quantile (qinvgauss) or random sample (rinvgauss) for the inverse Gaussian distribution with mean mu and inverse dispersion lambda. Elements of q or p that are missing will cause the corresponding elements of the result to be missing.

Author(s)

Gordon Smyth; Paul Bagshaw, Centre National d'Etudes des Telecommunications (DIH/DIPS), France (qinvgauss); Trevor Park, Department of Statistics, University of Florida

References

Chhikara, R. S., and Folks, J. Leroy, (1989). The inverse Gaussian distribution: Theory, methodology, and applications. Marcel Dekker, New York.

See Also

dinvGauss, pinvGauss, qinvGauss and rinvGauss in the SuppDists package.

Examples


y <- rinvgauss(20,1,2) # generate vector of 20 random numbers
p <- pinvgauss(y,1,2) # p should be uniform

[Package statmod version 1.3.8 Index]