power.fisher.test {statmod}	R Documentation

Power of Fisher's Exact Test for Comparing Proportions

Description

Calculate by simulation the power of Fisher's exact test for comparing two proportions given two margin counts.

Usage


power.fisher.test(p1, p2, n1, n2, alpha=0.05, nsim=100)

Arguments

p1 	 first proportion to be compared

p2 	 second proportion to be compared

n1 	 first sample size

n2 	 second sample size

alpha 	 significance level

nsim 	 number of simulated data sets

Details

Computes the power of Fisher's exact test for testing the null hypothesis that p1 equals p2 against the alternative that they are not equal.

Value

Estimated power of the test

Author(s)

Gordon Smyth

See Also

fisher.test, power.t.test

Examples


power.fisher.test(0.5,0.9,20,20) # 70

[Package statmod version 1.3.8 Index]