statmod-package {statmod}	R Documentation

Introduction to the StatMod Package

Description

This package includes a variety of functions for numerical analysis and statistical modelling. The functions are briefly summarized by type of application below.

Generalized Linear Models

The function tweedie defines a large class of generalized linear model families with power variance functions. It used in conjunction with the glm function, and widens the class of families that can be fitted.

qresiduals implements randomized quantile residuals for generalized linear models.

The functions canonic.digamma, unitdeviance.digamma, varfun.digamma, cumulant.digamma, d2cumulant.digamma, meanval.digamma and logmdigamma are used to fit double-generalized models, in which a link-linear model is fitted to the dispersion as well as to the mean. Spefically they are used to fit the dispersion submodel associated with a gamma glm.

Growth Curves

compareGrowthCurves, compareTwoGrowthCurves and meanT are functions to test for differences between growth curves with repeated measurements on subjects.

Limiting Dilution Analysis

The limdil function is used in the analysis of stem cell frequencies. It implements limiting dilution analysis using complemenary log-log binomial generalized linear model regression, with some improvements on previous programs.

Probability Distributions

The functions qinvgauss, dinvgauss, pinvgauss and rinvgauss provide probability calculations for the inverse Gaussian distribution.

gauss.quad and gauss.quad.prob compute Gaussian Quadrature with probability distributions.

Tests

hommel.test performs Hommel's multiple comparison tests.

power.fisher.test computes the power of Fisher's Exact Test for comparing proportions.

sage.test is a fast approximation to Fisher's exact test for each tag for comparing two Serial Analysis of Gene Expression (SAGE) libraries.

permp computes p-values for permutation tests when the permutations are randomly drawn.

Variance Models

mixedModel2, mixedModel2Fit and glmgam.fit fit mixed linear models.

remlscore and remlscoregamma fit heteroscedastic and varying dispersion models by REML. welding is an example data set.

Matrix Computations

matvec and vecmat facilitate multiplying matrices by vectors.

Author(s)

Gordon Smyth

[Package statmod version 1.4.36 Index]